pyproject.toml
[project]
name = "rpi-seism-common"
version = "1.1.0"
description = "Single source of truth for RPi Seismology schema"
readme = "README.md"
authors = [
    { name = "ch3p4ll3", email = "slinucs@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pyaml>=26.2.1",
    "pydantic>=2.12.5",
    "pydantic-extra-types>=2.11.1",
]

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"
